test(examples): move `common` files to `examples`; add tests for Identity and List
