+++ test/quasi.js
@@ -2,29 +2,25 @@ const { test } = require('./lib')
 
 const { of, empty, chainRec } = require('../src/quasi.js')
 
-const Identity = require('./common/Identity.js')
-const List = require('./common/List.js')
-// const Pair = require('./common/Pair.js')
-// const Func = require('./common/Func.js')
-// const Task = require('./common/Task.js')
-
-test('eq', t => {
-  const a = 'a'
-  const as = List.Cons(a, List.Nil)
+test('throws', t => {
   t.notThrow(() => {
     t.eqFL(of(0), chainRec((next, done, v) => of(v === 0 ? done(v) : next(v - 1)), 100000))
   }, 'chainRec is stacksafe')
   t.throws(() => { empty.map(a => a) }, 'throws on method calls which need value', TypeError)
   t.throws(() => { empty.ap(a => a) }, 'throws on method calls which need value', TypeError)
   t.throws(() => { empty.chain(a => a) }, 'throws on method calls which need value', TypeError)
-  t.throws(() => { empty.reduce((acc, a) => List.Cons(a, acc), List.Nil) }, 'throws on method calls which need value', TypeError)
+  t.throws(() => { empty.reduce((acc, a) => acc.concat([a]), []) }, 'throws on method calls which need value', TypeError)
   t.throws(() => { empty.traverse(a => a) }, 'throws on method calls which need value', TypeError)
   t.throws(() => { empty.extract(a => a) }, 'throws on method calls which need value', TypeError)
-  t.eqFL(Identity(0), chainRec((next, done, v) => Identity(v === 0 ? done(v) : next(v - 1)), 10))
-  t.eqFL(Identity(10), chainRec((_, done, v) => Identity(done(v)), 10))
+
+  t.end()
+})
+
+test('eq', t => {
+  const a = 'a'
+
   t.eqFL('<of>(a)', of(a).toString())
   t.eqFL('<of>(<empty>)', of(empty).toString())
-  t.ok(of(a).equals(Identity.of(a)))
   t.ok(of(empty).equals(empty))
   t.ok(of(empty).equals(of(empty)))
   t.ok(empty.equals(of(empty)))
@@ -32,24 +28,11 @@ test('eq', t => {
   t.notOk(of(a).equals(a))
   t.eqFL(of(empty), of(empty).concat(of(empty)))
   t.eqFL(of(a), of(a).constructor.of(a))
-  t.eqFL(Identity(of(a)), of(a).traverse((a) => Identity(a), Identity.of))
-  t.eqFL(of(Identity(of(a))), of(a).extend((a) => Identity(a)))
-  t.eqFL(of(Identity(of(empty))), empty.extend((a) => Identity(a)))
-  t.eqFL(Identity(a), of(Identity(a)).extract())
-  t.eqFL(of(List.Cons(a, List.Nil)), of(a).reduce((acc, a) => List.Cons(a, acc), List.Nil))
   t.eqFL(of(empty), empty.concat(of(empty)))
   t.eqFL(of(empty), of(empty).concat(empty))
-  t.eqFL(Identity(as), of(as).concat(Identity(List.Nil)))
-  t.eqFL(Identity(a), Identity(empty).concat(of(a)))
-  t.eqFL(of(Identity(a)), of(a).map(a => Identity(a)))
   t.eqFL(of(a), of(a).chain(a => of(a).chain(of)))
   t.eqFL(of(a), of(a).chain(of).chain(of))
   t.eqFL(of(a), of(a).ap(of(a => a)))
-  t.eqFL(Identity(a), of(a).chain(a => Identity(a)))
-  t.eqFL(of(Identity(a)), of(a).ap(of(a => Identity(a))))
-  t.eqFL(Identity(a), of(a).ap(Identity(a => a)))
-  t.eqFL(Identity(a), of(a).ap(of(a => a)).chain(a => of(a)).chain(a => Identity(a)))
-  t.eqFL(Identity(a), of(a).ap(Identity(a => a)))
 
   t.end()
 })
